Two different main.ptx versions main_doenet and main_virginia
They seem to use the same code eventually to reference a doenet file,
but the one where the doenet is in the main_doenet itself works. In
main_virginia it is referencing another pretext file and eventaully
has very simlilar tags to try to embedd doenet in a section, but it
fails with this error.

critical: Namespace prefix xi on include is not defined, line 28, column 70 (sec-1-1-Multiplying_Polynomials-the_FOIL_Method.ptx, line 28)